Completed basic product and ProductList componenets capable of making API requests and rendering props

+++ database/schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE products (
   image_one_url text,
   image_two_url text,
   page_url text,
-  price int,
+  price double,
   hearted boolean,
   brief_description text,
   collection_name varchar(255)
